backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.4.0,<2.10.0
alembic>=1.13.0
